+++ src/execution/operator/persistent/physical_insert.cpp
@@ -37,20 +37,20 @@ void PhysicalInsert::GetChunkInternal(ClientContext &context, DataChunk &chunk, 
 			ExpressionExecutor executor(chunk);
 			if (column_index_map.size() > 0) {
 				// columns specified by the user, use column_index_map
-				for (size_t i = 0; i < table->columns.size(); i++) {
+				for (uint64_t i = 0; i < table->columns.size(); i++) {
 					if (column_index_map[i] < 0) {
 						// insert default value
 						executor.ExecuteExpression(*table->bound_defaults[i], insert_chunk.data[i]);
 					} else {
 						// get value from child chunk
-						assert((size_t)column_index_map[i] < chunk.column_count);
+						assert((uint64_t)column_index_map[i] < chunk.column_count);
 						assert(insert_chunk.data[i].type == chunk.data[column_index_map[i]].type);
 						insert_chunk.data[i].Reference(chunk.data[column_index_map[i]]);
 					}
 				}
 			} else {
 				// no columns specified, just append directly
-				for (size_t i = 0; i < insert_chunk.column_count; i++) {
+				for (uint64_t i = 0; i < insert_chunk.column_count; i++) {
 					assert(insert_chunk.data[i].type == chunk.data[i].type);
 					insert_chunk.data[i].Reference(chunk.data[i]);
 				}
@@ -72,7 +72,7 @@ void PhysicalInsert::GetChunkInternal(ClientContext &context, DataChunk &chunk, 
 		for (auto &list : insert_values) {
 			if (column_index_map.size() > 0) {
 				// columns specified by the user, use column_index_map
-				for (size_t i = 0; i < table->columns.size(); i++) {
+				for (uint64_t i = 0; i < table->columns.size(); i++) {
 					if (column_index_map[i] < 0) {
 						// insert default value
 						executor.ExecuteExpression(*table->bound_defaults[i], temp_chunk.data[i]);
@@ -88,7 +88,7 @@ void PhysicalInsert::GetChunkInternal(ClientContext &context, DataChunk &chunk, 
 				}
 			} else {
 				// no columns specified
-				for (size_t i = 0; i < list.size(); i++) {
+				for (uint64_t i = 0; i < list.size(); i++) {
 					// execute the expressions to get the values
 					auto &expr = list[i];
 					if (expr->type == ExpressionType::VALUE_DEFAULT) {
