+++ src/execution/physical_plan/plan_distinct.cpp
@@ -11,7 +11,7 @@ unique_ptr<PhysicalOperator> PhysicalPlanGenerator::CreateDistinct(unique_ptr<Ph
 	// create a PhysicalHashAggregate that groups by the input columns
 	auto &types = child->GetTypes();
 	vector<unique_ptr<Expression>> groups, expressions;
-	for (size_t i = 0; i < types.size(); i++) {
+	for (uint64_t i = 0; i < types.size(); i++) {
 		groups.push_back(make_unique<BoundReferenceExpression>(types[i], i));
 	}
 	auto groupby =
